+++ src/main/java/fastui/layout/Layout.java
@@ -1,0 +1,7 @@
+package fastui.layout;
+
+import fastui.component.Component;
+
+public interface Layout {
+    void apply(Component parent);
+}